add tile sets to data base
