+++ packages/web-frontend/src/selectors/projectSelectors.js
@@ -9,6 +9,7 @@ import { createSelector } from 'reselect';
 import { DEFAULT_BOUNDS } from '../defaults';
 import { getLocationComponentValue, URL_COMPONENTS } from '../historyNavigation';
 import { selectLocation } from './utils';
+import { mapBoxToken } from '../utils';
 
 const selectAllProjects = state => state.project.projects;
 
@@ -41,3 +42,79 @@ export const selectAdjustedProjectBounds = createSelector(
     return project && project.bounds;
   },
 );
+
+// MAPBOX STYLES
+const GERRY_ACCESS_KEY =
+  'pk.eyJ1IjoiZ2VkY2tlbGx5IiwiYSI6ImNrY3BsZ2RwYTB3N20yc3FyaTZlNzhzNDUifQ.N61FIOcE-3RTksi9Tlm5ow#10.25/17.9782/102.6277';
+const GERRY_USERNAME = 'gedckelly';
+
+const makeStyleUrl = ({ styleId, accessKey = GERRY_ACCESS_KEY, username = GERRY_USERNAME }) =>
+  `https://api.mapbox.com/styles/v1/${username}/${styleId}/tiles/256/{z}/{x}/{y}@2x?access_token=${accessKey}`;
+
+const TILE_SETS = [
+  {
+    key: 'osm',
+    label: 'Open Streets',
+    thumbnail: '/images/osm.png',
+    url: 'http://{s}.tile.osm.org/{z}/{x}/{y}.png',
+  },
+  {
+    key: 'satellite',
+    label: 'Satellite',
+    thumbnail: '/images/satellite.png',
+    url: `https://api.mapbox.com/styles/v1/mapbox/satellite-streets-v10/tiles/256/{z}/{x}/{y}?access_token=${mapBoxToken}`,
+  },
+  {
+    key: 'waterways',
+    label: 'Waterways',
+    thumbnail: '/images/waterways.png',
+    url: makeStyleUrl({ styleId: 'ckemdct811px619qklzgvvg53' }),
+  },
+  {
+    key: 'roads',
+    label: 'Roads',
+    thumbnail: '/images/roads.png',
+    url: makeStyleUrl({ styleId: 'ckenp4uq10dfq1anzert7iot7' }),
+    legendItems: [
+      {
+        color: '#D13333',
+        label: 'Ethnic group one',
+      },
+      {
+        color: '#E37F49',
+        label: 'Ethnic group two',
+      },
+      {
+        color: '#E12EC5',
+        label: 'Ethnic group three',
+      },
+      {
+        color: '#22D489',
+        label: 'Ethnic group four',
+      },
+      {
+        color: '#2196F3',
+        label: 'Ethnic group five',
+      },
+    ],
+  },
+  {
+    key: 'terrain',
+    label: 'Terrain',
+    thumbnail: '/images/terrain.png',
+    url: makeStyleUrl({ styleId: 'ckenu2thw0ibl1anzk5aarzu6' }),
+  },
+];
+
+export const selectTileSets = state => {
+  const currentProject = selectCurrentProject(state);
+  const tileSetKeys = currentProject.tileSets
+    ? currentProject.tileSets.split(',')
+    : ['osm', 'satellite'];
+  return TILE_SETS.filter(tileSet => tileSetKeys.includes(tileSet.key));
+};
+
+export const selectActiveTileSet = state => {
+  const tileSets = selectTileSets(state);
+  return tileSets.find(tileSet => tileSet.key === state.map.activeTileSetKey);
+};
